+++ lti_consumer/apps.py
@@ -26,4 +26,4 @@ class LTIConsumerApp(AppConfig):
 
     def ready(self):
         # pylint: disable=unused-import,import-outside-toplevel
-        import lti_consumer.signals
+        from lti_consumer.signals import signals