+++ lti_consumer/plugin/urls.py
@@ -10,7 +10,8 @@ from rest_framework import routers
 from lti_consumer.plugin.views import (LtiAgsLineItemViewset,  # LTI Advantage URLs; LTI NRPS URLs
                                        LtiNrpsContextMembershipViewSet, access_token_endpoint,
                                        deep_linking_content_endpoint, deep_linking_response_endpoint,
-                                       launch_gate_endpoint, public_keyset_endpoint)
+                                       launch_gate_endpoint, public_keyset_endpoint,
+                                       start_proctoring_assessment_endpoint)
 
 # LTI 1.3 APIs router
 router = routers.SimpleRouter(trailing_slash=False)
@@ -58,5 +59,10 @@ urlpatterns = [
     re_path(
         r'lti_consumer/v1/lti/(?P<lti_config_id>[-\w]+)/',
         include(router.urls)
+    ),
+    path(
+        'lti_consumer/v1/start_proctoring_assessment',
+        start_proctoring_assessment_endpoint,
+        name='lti_consumer.start_proctoring_assessment_endpoint'
     ),
 ]